Replace usage of expire time for cache expiration check
Use leases to check credential validity.

Due to Vault tokens hierarchy, a credential or token may be revoked even if
the expire time has not been reached, because a parent token has reached
the expire time or being manually revoked and thus revokes all children.

The only way to work with this is to check the associated lease each time.

> When a parent token is revoked, all of its child tokens -- and all of their leases -- are revoked as well.
https://developer.hashicorp.com/vault/docs/concepts/tokens#token-hierarchies-and-orphan-tokens